requirements.txt
# Dependencias del Running Coach Agent

# API de Anthropic Claude
anthropic>=0.39.0

# Cliente HTTP
requests>=2.31.0

# Variables de entorno desde .env
python-dotenv>=1.0.0

# Python 3.8+ (requerido)
